pkg/rpcserver: split rpcserver.go
Split out most of the Runner functionality into a separate file.

This should make it easier to reason about what rpcserver.go does and
it also makes further pkg/rpcserver refactoring simpler.
