+++ pkg/rpcserver/rpcserver.go
@@ -9,7 +9,6 @@ import (
 	"fmt"
 	"maps"
 	"math/rand"
-	"os"
 	"slices"
 	"sort"
 	"strings"
@@ -22,7 +21,6 @@ import (
 	"github.com/google/syzkaller/pkg/fuzzer/queue"
 	"github.com/google/syzkaller/pkg/log"
 	"github.com/google/syzkaller/pkg/mgrconfig"
-	"github.com/google/syzkaller/pkg/osutil"
 	"github.com/google/syzkaller/pkg/signal"
 	"github.com/google/syzkaller/pkg/stats"
 	"github.com/google/syzkaller/pkg/vminfo"
@@ -66,33 +64,13 @@ type Server struct {
 	canonicalModules *cover.Canonicalizer
 	coverFilter      []uint64
 
-	mu            sync.Mutex
-	runners       map[string]*Runner
-	info          map[string]VMState
-	execSource    queue.Source
-	triagedCorpus atomic.Bool
-
-	statExecRetries        *stats.Val
-	statExecutorRestarts   *stats.Val
-	statExecBufferTooSmall *stats.Val
-	statVMRestarts         *stats.Val
-	statNoExecRequests     *stats.Val
-	statNoExecDuration     *stats.Val
-}
-
-type Runner struct {
-	stopped       bool
-	finished      chan bool
-	injectExec    chan<- bool
-	infoc         chan chan []byte
-	conn          *flatrpc.Conn
-	machineInfo   []byte
-	canonicalizer *cover.CanonicalizerInstance
-	nextRequestID int64
-	requests      map[int64]*queue.Request
-	executing     map[int64]bool
-	lastExec      *LastExecuting
-	rnd           *rand.Rand
+	mu             sync.Mutex
+	runners        map[string]*Runner
+	info           map[string]VMState
+	execSource     queue.Source
+	triagedCorpus  atomic.Bool
+	statVMRestarts *stats.Val
+	*runnerStats
 }
 
 func New(cfg *mgrconfig.Config, mgr Manager, debug bool) (*Server, error) {
@@ -137,20 +115,23 @@ func newImpl(cfg *Config, mgr Manager) (*Server, error) {
 			stats.Console, stats.Rate{}, stats.Prometheus("syz_exec_total")),
 		StatNumFuzzing: stats.Create("fuzzing VMs", "Number of VMs that are currently fuzzing",
 			stats.Console, stats.Link("/vms")),
-		statExecRetries: stats.Create("exec retries",
-			"Number of times a test program was restarted because the first run failed",
-			stats.Rate{}, stats.Graph("executor")),
-		statExecutorRestarts: stats.Create("executor restarts",
-			"Number of times executor process was restarted", stats.Rate{}, stats.Graph("executor")),
-		statExecBufferTooSmall: stats.Create("buffer too small",
-			"Program serialization overflowed exec buffer", stats.NoGraph),
 		statVMRestarts: stats.Create("vm restarts", "Total number of VM starts",
 			stats.Rate{}, stats.NoGraph),
-		statNoExecRequests: stats.Create("no exec requests",
-			"Number of times fuzzer was stalled with no exec requests", stats.Rate{}),
-		statNoExecDuration: stats.Create("no exec duration",
-			"Total duration fuzzer was stalled with no exec requests (ns/sec)", stats.Rate{}),
+		runnerStats: &runnerStats{
+			statExecRetries: stats.Create("exec retries",
+				"Number of times a test program was restarted because the first run failed",
+				stats.Rate{}, stats.Graph("executor")),
+			statExecutorRestarts: stats.Create("executor restarts",
+				"Number of times executor process was restarted", stats.Rate{}, stats.Graph("executor")),
+			statExecBufferTooSmall: stats.Create("buffer too small",
+				"Program serialization overflowed exec buffer", stats.NoGraph),
+			statNoExecRequests: stats.Create("no exec requests",
+				"Number of times fuzzer was stalled with no exec requests", stats.Rate{}),
+			statNoExecDuration: stats.Create("no exec duration",
+				"Total duration fuzzer was stalled with no exec requests (ns/sec)", stats.Rate{}),
+		},
 	}
+	serv.runnerStats.statExecs = serv.StatExecs
 	s, err := flatrpc.ListenAndServe(cfg.RPC, serv.handleConn)
 	if err != nil {
 		return nil, err
@@ -205,19 +186,7 @@ func (serv *Server) RunnerStatus(name string) []byte {
 	if runner == nil {
 		return []byte("VM is not alive")
 	}
-	resc := make(chan []byte, 1)
-	timeout := time.After(time.Minute)
-	select {
-	case runner.infoc <- resc:
-	case <-timeout:
-		return []byte("VM loop is not responding")
-	}
-	select {
-	case res := <-resc:
-		return res
-	case <-timeout:
-		return []byte("VM is not responding")
-	}
+	return runner.queryStatus()
 }
 
 func (serv *Server) handleConn(conn *flatrpc.Conn) {
@@ -363,222 +332,8 @@ func (serv *Server) connectionLoop(runner *Runner) error {
 
 	serv.StatNumFuzzing.Add(1)
 	defer serv.StatNumFuzzing.Add(-1)
-	var infoc chan []byte
-	defer func() {
-		if infoc != nil {
-			infoc <- []byte("VM has crashed")
-		}
-	}()
-	for {
-		if infoc == nil {
-			select {
-			case infoc = <-runner.infoc:
-				msg := &flatrpc.HostMessage{
-					Msg: &flatrpc.HostMessages{
-						Type:  flatrpc.HostMessagesRawStateRequest,
-						Value: &flatrpc.StateRequest{},
-					},
-				}
-				if err := flatrpc.Send(runner.conn, msg); err != nil {
-					return err
-				}
-			default:
-			}
-		}
-		for len(runner.requests)-len(runner.executing) < 2*serv.cfg.Procs {
-			req := serv.execSource.Next()
-			if req == nil {
-				break
-			}
-			if err := serv.sendRequest(runner, req); err != nil {
-				return err
-			}
-		}
-		if len(runner.requests) == 0 {
-			// The runner has not requests at all, so don't wait to receive anything from it.
-			time.Sleep(10 * time.Millisecond)
-			continue
-		}
-		raw, err := flatrpc.Recv[*flatrpc.ExecutorMessageRaw](runner.conn)
-		if err != nil {
-			return err
-		}
-		if raw.Msg == nil || raw.Msg.Value == nil {
-			return errors.New("received no message")
-		}
-		switch msg := raw.Msg.Value.(type) {
-		case *flatrpc.ExecutingMessage:
-			err = serv.handleExecutingMessage(runner, msg)
-		case *flatrpc.ExecResult:
-			err = serv.handleExecResult(runner, msg)
-		case *flatrpc.StateResult:
-			if infoc != nil {
-				buf := new(bytes.Buffer)
-				fmt.Fprintf(buf, "pending requests on the VM:")
-				for id := range runner.requests {
-					fmt.Fprintf(buf, " %v", id)
-				}
-				fmt.Fprintf(buf, "\n\n")
-				infoc <- append(buf.Bytes(), msg.Data...)
-				infoc = nil
-			}
-		default:
-			return fmt.Errorf("received unknown message type %T", msg)
-		}
-		if err != nil {
-			return err
-		}
-	}
-}
 
-func (serv *Server) sendRequest(runner *Runner, req *queue.Request) error {
-	if err := req.Validate(); err != nil {
-		panic(err)
-	}
-	runner.nextRequestID++
-	id := runner.nextRequestID
-	var flags flatrpc.RequestFlag
-	if req.ReturnOutput {
-		flags |= flatrpc.RequestFlagReturnOutput
-	}
-	if req.ReturnError {
-		flags |= flatrpc.RequestFlagReturnError
-	}
-	allSignal := make([]int32, len(req.ReturnAllSignal))
-	for i, call := range req.ReturnAllSignal {
-		allSignal[i] = int32(call)
-	}
-	// Do not let too much state accumulate.
-	const restartIn = 600
-	resetFlags := flatrpc.ExecFlagCollectSignal | flatrpc.ExecFlagCollectCover | flatrpc.ExecFlagCollectComps
-	opts := req.ExecOpts
-	if req.ExecOpts.ExecFlags&resetFlags != 0 && runner.rnd.Intn(restartIn) == 0 {
-		opts.EnvFlags |= flatrpc.ExecEnvResetState
-	}
-	if serv.cfg.Debug {
-		opts.EnvFlags |= flatrpc.ExecEnvDebug
-	}
-	var data []byte
-	if req.BinaryFile == "" {
-		progData, err := req.Prog.SerializeForExec()
-		if err != nil {
-			// It's bad if we systematically fail to serialize programs,
-			// but so far we don't have a better handling than counting this.
-			// This error is observed a lot on the seeded syz_mount_image calls.
-			serv.statExecBufferTooSmall.Add(1)
-			req.Done(&queue.Result{Status: queue.ExecFailure})
-			return nil
-		}
-		data = progData
-	} else {
-		flags |= flatrpc.RequestFlagIsBinary
-		fileData, err := os.ReadFile(req.BinaryFile)
-		if err != nil {
-			req.Done(&queue.Result{
-				Status: queue.ExecFailure,
-				Err:    err,
-			})
-			return nil
-		}
-		data = fileData
-	}
-	msg := &flatrpc.HostMessage{
-		Msg: &flatrpc.HostMessages{
-			Type: flatrpc.HostMessagesRawExecRequest,
-			Value: &flatrpc.ExecRequest{
-				Id:        id,
-				ProgData:  data,
-				Flags:     flags,
-				ExecOpts:  &opts,
-				AllSignal: allSignal,
-			},
-		},
-	}
-	runner.requests[id] = req
-	return flatrpc.Send(runner.conn, msg)
-}
-
-func (serv *Server) handleExecutingMessage(runner *Runner, msg *flatrpc.ExecutingMessage) error {
-	req := runner.requests[msg.Id]
-	if req == nil {
-		return fmt.Errorf("can't find executing request %v", msg.Id)
-	}
-	proc := int(msg.ProcId)
-	if proc < 0 || proc >= serv.cfg.Procs {
-		return fmt.Errorf("got bad proc id %v", proc)
-	}
-	serv.StatExecs.Add(1)
-	if msg.Try == 0 {
-		if msg.WaitDuration != 0 {
-			serv.statNoExecRequests.Add(1)
-			// Cap wait duration to 1 second to avoid extreme peaks on the graph
-			// which make it impossible to see real data (the rest becomes a flat line).
-			serv.statNoExecDuration.Add(int(min(msg.WaitDuration, 1e9)))
-		}
-	} else {
-		serv.statExecRetries.Add(1)
-	}
-	runner.lastExec.Note(int(msg.Id), proc, req.Prog.Serialize(), osutil.MonotonicNano())
-	select {
-	case runner.injectExec <- true:
-	default:
-	}
-	runner.executing[msg.Id] = true
-	return nil
-}
-
-func (serv *Server) handleExecResult(runner *Runner, msg *flatrpc.ExecResult) error {
-	req := runner.requests[msg.Id]
-	if req == nil {
-		return fmt.Errorf("can't find executed request %v", msg.Id)
-	}
-	delete(runner.requests, msg.Id)
-	delete(runner.executing, msg.Id)
-	if msg.Info != nil {
-		for len(msg.Info.Calls) < len(req.Prog.Calls) {
-			msg.Info.Calls = append(msg.Info.Calls, &flatrpc.CallInfo{
-				Error: 999,
-			})
-		}
-		msg.Info.Calls = msg.Info.Calls[:len(req.Prog.Calls)]
-		if msg.Info.Freshness == 0 {
-			serv.statExecutorRestarts.Add(1)
-		}
-		if !serv.cfg.Cover && req.ExecOpts.ExecFlags&flatrpc.ExecFlagCollectSignal != 0 {
-			// Coverage collection is disabled, but signal was requested => use a substitute signal.
-			addFallbackSignal(req.Prog, msg.Info)
-		}
-		for i := 0; i < len(msg.Info.Calls); i++ {
-			call := msg.Info.Calls[i]
-			call.Cover = runner.canonicalizer.Canonicalize(call.Cover)
-			call.Signal = runner.canonicalizer.Canonicalize(call.Signal)
-		}
-		if len(msg.Info.ExtraRaw) != 0 {
-			msg.Info.Extra = msg.Info.ExtraRaw[0]
-			for _, info := range msg.Info.ExtraRaw[1:] {
-				// All processing in the fuzzer later will convert signal/cover to maps and dedup,
-				// so there is little point in deduping here.
-				msg.Info.Extra.Cover = append(msg.Info.Extra.Cover, info.Cover...)
-				msg.Info.Extra.Signal = append(msg.Info.Extra.Signal, info.Signal...)
-			}
-			msg.Info.Extra.Cover = runner.canonicalizer.Canonicalize(msg.Info.Extra.Cover)
-			msg.Info.Extra.Signal = runner.canonicalizer.Canonicalize(msg.Info.Extra.Signal)
-			msg.Info.ExtraRaw = nil
-		}
-	}
-	status := queue.Success
-	var resErr error
-	if msg.Error != "" {
-		status = queue.ExecFailure
-		resErr = errors.New(msg.Error)
-	}
-	req.Done(&queue.Result{
-		Status: status,
-		Info:   msg.Info,
-		Output: slices.Clone(msg.Output),
-		Err:    resErr,
-	})
-	return nil
+	return runner.connectionLoop()
 }
 
 func checkRevisions(a *flatrpc.ConnectRequest, target *prog.Target) {
@@ -666,6 +421,9 @@ func (serv *Server) printMachineCheck(checkFilesInfo []*flatrpc.FileInfo, enable
 
 func (serv *Server) CreateInstance(name string, injectExec chan<- bool) {
 	runner := &Runner{
+		source:     serv.baseSource,
+		cover:      serv.cfg.Cover,
+		debug:      serv.cfg.Debug,
 		injectExec: injectExec,
 		infoc:      make(chan chan []byte),
 		finished:   make(chan bool),
@@ -673,6 +431,8 @@ func (serv *Server) CreateInstance(name string, injectExec chan<- bool) {
 		executing:  make(map[int64]bool),
 		lastExec:   MakeLastExecuting(serv.cfg.Procs, 6),
 		rnd:        rand.New(rand.NewSource(time.Now().UnixNano())),
+		stats:      serv.runnerStats,
+		procs:      serv.cfg.Procs,
 	}
 	serv.mu.Lock()
 	if serv.runners[name] != nil {
@@ -703,19 +463,7 @@ func (serv *Server) ShutdownInstance(name string, crashed bool) ([]ExecRecord, [
 	delete(serv.runners, name)
 	serv.info[name] = VMState{StateOffline, time.Now()}
 	serv.mu.Unlock()
-	if runner.conn != nil {
-		// Wait for the connection goroutine to finish and stop touching data.
-		// If conn is nil before we removed the runner, then it won't touch anything.
-		<-runner.finished
-	}
-	for id, req := range runner.requests {
-		status := queue.Restarted
-		if crashed && runner.executing[id] {
-			status = queue.Crashed
-		}
-		req.Done(&queue.Result{Status: status})
-	}
-	return runner.lastExec.Collect(), runner.machineInfo
+	return runner.shutdown(crashed)
 }
 
 func (serv *Server) DistributeSignalDelta(plus, minus signal.Signal) {
@@ -726,34 +474,11 @@ func (serv *Server) DistributeSignalDelta(plus, minus signal.Signal) {
 	})
 }
 
-func (runner *Runner) sendSignalUpdate(plus, minus []uint64) error {
-	msg := &flatrpc.HostMessage{
-		Msg: &flatrpc.HostMessages{
-			Type: flatrpc.HostMessagesRawSignalUpdate,
-			Value: &flatrpc.SignalUpdate{
-				NewMax:  runner.canonicalizer.Decanonicalize(plus),
-				DropMax: runner.canonicalizer.Decanonicalize(minus),
-			},
-		},
-	}
-	return flatrpc.Send(runner.conn, msg)
-}
-
 func (serv *Server) TriagedCorpus() {
 	serv.triagedCorpus.Store(true)
 	serv.foreachRunnerAsync(func(runner *Runner) {
 		runner.sendStartLeakChecks()
 	})
-}
-
-func (runner *Runner) sendStartLeakChecks() error {
-	msg := &flatrpc.HostMessage{
-		Msg: &flatrpc.HostMessages{
-			Type:  flatrpc.HostMessagesRawStartLeakChecks,
-			Value: &flatrpc.StartLeakChecks{},
-		},
-	}
-	return flatrpc.Send(runner.conn, msg)
 }
 
 // foreachRunnerAsync runs callback fn for each connected runner asynchronously.
@@ -766,28 +491,5 @@ func (serv *Server) foreachRunnerAsync(fn func(runner *Runner)) {
 		if runner.conn != nil {
 			go fn(runner)
 		}
-	}
-}
-
-// addFallbackSignal computes simple fallback signal in cases we don't have real coverage signal.
-// We use syscall number or-ed with returned errno value as signal.
-// At least this gives us all combinations of syscall+errno.
-func addFallbackSignal(p *prog.Prog, info *flatrpc.ProgInfo) {
-	callInfos := make([]prog.CallInfo, len(info.Calls))
-	for i, inf := range info.Calls {
-		if inf.Flags&flatrpc.CallFlagExecuted != 0 {
-			callInfos[i].Flags |= prog.CallExecuted
-		}
-		if inf.Flags&flatrpc.CallFlagFinished != 0 {
-			callInfos[i].Flags |= prog.CallFinished
-		}
-		if inf.Flags&flatrpc.CallFlagBlocked != 0 {
-			callInfos[i].Flags |= prog.CallBlocked
-		}
-		callInfos[i].Errno = int(inf.Error)
-	}
-	p.FallbackSignal(callInfos)
-	for i, inf := range callInfos {
-		info.Calls[i].Signal = inf.Signal
 	}
 }
